Make tickers great again
