+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__
 __pycache__/*
 *.pyc
+CREDENTIALS.json
