+++ CREDENTIALS.json
@@ -1,0 +1,6 @@
+{
+    "screener": {
+        "username": "chinmaybuys@gmail.com",
+        "password": "mKa-s-sr2-r-C3K"
+    }
+}
